Switched Test to use TestProvideer
